backend/requirements.txt
Flask==3.1.2
Flask-Cors==6.0.1
python-dotenv==1.0.1
supabase==2.4.3
requests==2.32.3
gunicorn


